requirements.txt
Flask==2.3.3
Pillow==10.0.0
requests==2.31.0
Werkzeug==2.3.7
gunicorn==21.2.0